# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Departamento de Sistemas",
        "description": "Documentación de la API para el sistema de gestión del Departamento de Sistemas.",
        "contact": {
            "email": "soporte@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/cp-dependencias": {
            "get": {
                "tags": [
                    "CP Dependencias"
                ],
                "summary": "Listar dependencias",
                "description": "Obtiene la lista de dependencias. Requiere permiso cp_dependencia.read.",
                "operationId": "a41556a014f9f7c96357bc0cab7edf09",
                "responses": {
                    "200": {
                        "description": "Lista de dependencias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CP Dependencias"
                ],
                "summary": "Crear dependencia",
                "description": "Crea una nueva dependencia. Requiere permiso cp_dependencia.create.",
                "operationId": "9ca688b3f54e1521bccd546f3e16b518",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codigo": {
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Dirección General"
                                    },
                                    "sede_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dependencia creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cp-dependencias/{id}": {
            "get": {
                "tags": [
                    "CP Dependencias"
                ],
                "summary": "Obtener dependencia",
                "description": "Obtiene los detalles de una dependencia. Requiere permiso cp_dependencia.read.",
                "operationId": "a24915b848bd362fce73d3e2c76689fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la dependencia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dependencia no encontrada"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CP Dependencias"
                ],
                "summary": "Actualizar dependencia",
                "description": "Actualiza una dependencia existente. Requiere permiso cp_dependencia.update.",
                "operationId": "0f3b09198362f30b2285de28dce3ddb6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codigo": {
                                        "type": "integer",
                                        "example": 124
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Dirección General Actualizada"
                                    },
                                    "sede_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dependencia actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dependencia no encontrada"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CP Dependencias"
                ],
                "summary": "Eliminar dependencia",
                "description": "Elimina una dependencia. Requiere permiso cp_dependencia.delete.",
                "operationId": "ff45f7cdde7406543f79a5650c59ac1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dependencia eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dependencia no encontrada"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "title": "Respuesta Estándar",
                "description": "Estructura de respuesta estándar para la API",
                "properties": {
                    "mensaje": {
                        "type": "string",
                        "example": "Operación exitosa"
                    },
                    "objeto": {
                        "type": "object",
                        "example": []
                    },
                    "status": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT en el formato: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}